Fixes some of the camera positioning issues.
Will restore the camera position on page refresh, or when switching
between tabs. Restoration on page refresh only works when the tabs have
been switched once, do to the fact that we only set camera position when
the tab is unloaded and not on every camera change. This will probably
need to be fixed.