Fixes dup. issue when switching skeleton tabs.
The skeleton viewer tabs were not updating the content when switching
between them directly. This was because there was no code to check to
see if the query props had changed and then delete bodies or
compartments that had disappeared from the query. The result was that
they would all get loaded into all the tabs.