Adds processDownload() to Autapses plugin.
This function is new and can be used by plugins to control the output of
the download button in the explorer. By default they will just dump the
response from the Neo4j server, but that might not be desired if there
is a lot of post processing to generate the data shown in the browser.
This gives the plugin developer a chance to create their own download
function to format the csv the way they want to.