Restores PartnerCompletenessView to query plugin.
The PartnerCompleteness query plugin was accidentally converted to
use the SimpleTable view. This restores it to use the view it was
supposed to use. The PartnerCompletenessView also had to be completely
rewritten to work with the new url system. The original version was
storing all the table data in visProps, which would have ended up in the
url and broken it. Moved everything to state, except the filter value
and the status selection.