Removes simple connections code from FindNeurons query
This code was failing and causing the site to crash, and I can't see why
it was there in the first place. Removing it makes the site show the
single neuron that was requested and clicking on the pre and post
columns of the table take the user to the simple connections results.