View plugins import now prevents neuroglancer crash
The automatic import of all view plugins caused the parent application
to crash in specific browsers, because the neuroglancer plugin executed
code relying on webGL2, which wasn't present. This update adds a webgl2
check, before importing the neuroglancer code, so that it wont execute
the offending code and crash the application.