Fixes column sorting in SimpleTable
The column sorting was thrown off when hiding columns, because it was
relying on an index that was being set after the columns were removed.
This meant that the index pointed to the wrong column when the sort
executed. The trick is to apply the index and then filter out the
columns that are hidden.