Fixes colorbox layout when used in miniroibargraph
There were cases where the colorbox would get rendered in the wrong
place if it contained no text. This 'hack' adds a '.' to the div that is
the same color as the background. This causes the browser rendering
engine to treat it the same way as the div with text in it and render it
the way we want it to be rendered.