Fixes clone issue in production.
Clone is set to a depth of 2. Anything less than that can result in an

Uncaught TypeError: Illegal

It looks like the clone code works fine in development, but when
transpiled for production it throws the above error, because it
can't clone an HTMLElement. This fix only works if we don't put an HTML
element higher up in the object. The ultimate fix is to remove the
results key from the query object. That will take some substantial
plumbing.