Adds new heatmap code to change data display.
null values are now greyed out to distinguish them from the 0 values. In
addition, the median row forces the drawing of the labels, so that the
numbers can always bee seen, no matter how small the block is. Finally
the column headers have had their index added, so that they don't
overwrite each other, when they are not unique.