Adds colored labels and disables maxWidth on heatmap.
The labels on the celltype heatmaps are now colored red if they are not
a reference sequence. Also, the heatmap width is no longer limited to
the width of the parent container, so that it can be scrolled
horizontally when it becomes too long, as opposed to it being shrunk to
fit.