Fixes NeuronInputField for SimpleConntections & FindNeurons
Both of these originally used a regular expression to allow for complex
searches. The input now searches the database for the string supplied to
autocomplete the input. In order to do this we have to add the
'enable_contains: true' parameter to the query or it will fail when
characters such as '(' or '?' are used.