Show more info for each option in NeuronInputField
There are instances where the label displayed doesn't appear to match
the search term. This is because it matches the type, body id or
instance associated with the label. This was causing some confusion, so
the other parts of the search are tacked onto the end of the label to
try to show why it was selected as an autocomplete option.