Adds localStorage check to FindNeurons
It is a pain in the neck to constantly have to toggle the advanced
search on if you are a power user. This sets the value in localStorage,
so that once it is set, it doesn't need to be changed until the browser
cache is cleared or the individual logs into another browser.