Fixes duplicate entries NeuronInputField
The original code in NeuronInputField used a set to make sure all the
values were unique. When the code was updated to pass a label and a value
as an array, those array objects were all considered unique, so the Set
was not working as intended. Moving the labels into a lookup hash meant
that the values could be passed and saved as unique strings as intended
and the duplicate values would disappear. Then the label could be
applied from the lookup later.