Changes arguments to processResults()
The arguments list to the procesingResults function got too long, so it
has been changed to accept an object of named arguments. These can then
be unpacked as needed in the function.

This also allowed us to pass in the roinfo to the
ROIsIntersectingNeurons component so that ROI abbreviation mouse overs
will work.