Fixes rounding errors with the roi bragraphs.
There were a number of edge cases where the rounding would be off by
one. This was caused by the code simply rounding the percentage values
of each ROI. This can lead to a rounding error when too many of the ROIs
are rounded up. Had to switch to a more complicated method that was
described in the answer to:

https://stackoverflow.com/qeuestions/13483430/how-to-make-rounded-percentages-add-up-to-100

https://en.wikipedia.org/wiki/Largest_remainder_method