Adds error handling for bad or missing data.
IF the total length of ROIs passed in is 0 then the code bails early and
shows a greyed out box with 0% in it. If the numbers provided from the
database are incorrect then the rounding code will fail after
calculating the percentages. This error is trapped and an error box is
displayed to indicate that the percentage calculation didn't work.