Calculate 'None' ROI on the client.
We used to only calculate the 'None' ROI when one wasn't present in the
data returned from the server. The server was returning a value for None
that didn't match what we were expecting. Now we calculate None for all
returned values on the client, to make sure it is correct for display.