Fixes NeuroglancerViewer layer import problems.
There was no way to import more than one segmentation layer. This was
fixed by changing the code that modifies the segmntation layer to be
more specific and only modify the layer that has a dataInstance of
segmentation. The visible setting was being ignored when set to false,
because the check was checking for a return value from the key, which
was false when set to false. Now the code simply gets the value and if
missing returns a default of true.