SunburstFormatter: adds guard when adding ROIs
As the code loops over the rawData it adds the roiData.post value to the
total value for each ROI. In some cases that value was undefined, which
lead to the total value becoming NaN. This resulted in the removal of
the neuron type from the sunburst result. Now it logs a warning, but
keeps the total value for the ones that do have real numbers.