Moves FindNeurons cypher query into client plugin.
The original cypher generation code is in the neuprintHTTP server, which
makes it difficult to make updates to the cypher to get new columns into
the results table. Moving the cypher building here makes it possible to
update the cypher at the same time we update the table columns, without
needing to redeploy the server code.