Updates FindNeuron results Processing
Simplifies the way the data is fetched and processed by using the json
objects returned from a new cypher query, instead of looping over an
array and trying to place the data at the correct array index. This also
removes a lot of redundant column accesses that were a source of errors
and maintenance issues.