Changes Motif to VimoMotif to change order in webpage.
The Query menu is sorted alphabetically by the name of the files on
disk. In order to get the Motif Query to go to the bottom of the top
level search types, the name needs to start with a letter after 'S'.
Changing it to VimoMotif.jsx achieves the goal.