fix: Changes find neurons cypher to = from CONTAINS
The CONTAINS query was too generic and was returning Mi18 types and
others, when the user selected Mi1. This cypher change makes sure that
only the type, instance or bodyid that is selected in the autocomplete
result is returned. Individuals that wish to return all types that start
with a specific string, eg: Mi1, can use the Advanced Search interface,
where they can use regular expressions to achieve the search.