feat: Changes CellTypeHeatMap to canvas & not d3
There were a number of cases where the SVG generated by d3 contained too
many nodes to render on a high spec laptop. To fix this the heatmap code
was converted to a canvas, so that only one DOM element is added to the
page, instead of the potentially 200 - 300 thousand nodes from an SVG.
This also adds a pop up modal that shows the values for the square under
the pointer.