feat: Removes some default columns from FindNeurons
This removes some of the default columns that were added to the
FindNeurons plugin, so that they can be added on a per dataset basis.
A number of these columns were not needed in older datasets and added
unnecessary clutter to the UI. If someone does want to add these columns
for a specific dataset, they can add them via the n.neuronColumns meta
value in the neo4j database for that dataset.