feat: Adds links to FindNeurons type columns
This allows the person sending data back from neuprintHTTP to change the
type column data into a JSON string, which contains an href and a label
key. If that is found, a link to the href will be created. If the JSON
string is not found, then the default string value of the type will be
displayed.