fix: Adds help message for 1 neuron in CellTypeView
There are instances where the selected cell type has only one associated
neuron in the cell type view. When that happens, the neuprintHTTP server
does not send back for the missing or common, inputs and outputs. This
was leaving the space on the page blank with no explanation as to why.
Now the site displays some text to explain that the data is not
calculated for single neuron cell types.