fix: Find Neurons download function restored
The download button on the find neurons results was not working, because
it was expecting a different data structure. The code would throw an
error when trying to parse and item that was not a serialized JSON
object. This updates the processDownload function to use the new data
structure and create the download csv.