feat: Adds "after" to enable column sorting.
The column data that is sent from the neuprint server via the
neuronColumns meta information can now include an "after" key. If this
key contains an id of another column, then the column will be moved
after the column identified. This makes it possible to change the column
order on a per dataset basis.